// pronti-frontend-main/front/src/app/modulos/configuracion-general/form/form-configuracion-general.enum.ts
export enum FormConfiguracionGeneralEnum {
  // nombre = 'nombre',
  ivaPorcentaje = 'ivaPorcentaje',
  tasaCambioContrato = 'tasaCambioContrato',
  minCuotaMoraRefinanciamiento = 'minCuotaMoraRefinanciamiento',
  maxContratosEnGrupo = 'maxContratosEnGrupo',
  cuotaAdministrativa = 'cuotaAdministrativa',
  aportePatronalIess = 'aportePatronalIess',
  aportePersonalIess = 'aportePersonalIess',
  numDiasVacacionesAlAnio = 'numDiasVacacionesAlAnio',
  sueldoBasico = 'sueldoBasico',
  tasaCargoAdjudicacion = 'tasaCargoAdjudicacion',
  rastreo = 'rastreo',
  dispositivo = 'dispositivo',
  asumeIva='asumeIva',
  elaboradorRol = 'elaboradorRol',
  revisadorRol = 'revisadorRol'
}